third_party/WebKit/LayoutTests/bluetooth/requestDevice/canonicalizeFilter/device-name-longer-than-29-bytes.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
bluetooth_test(() => {
  const DEVICE_NAME = 'a_device_name_that_is_longer_than_29_bytes_but_shorter_than_248_bytes';

  return setUpPreconnectedDevice({name: DEVICE_NAME})
    .then(() => requestDeviceWithKeyDown({ filters: [{name: DEVICE_NAME}]}))
    .then(device => assert_equals(device.name, DEVICE_NAME));
}, 'A device name between 29 and 248 bytes is valid.');
</script>
